requirements.txt
schedule
python-telegram-bot
pytz
python-dotenv
tzdata